+++ Source/MeshIO/imstkAssimpMeshIO.cpp
@@ -112,8 +112,8 @@ AssimpMeshIO::convertAssimpMesh(aiMesh* importedMesh)
     // Vertex normals, tangents, and bitangents
     std::shared_ptr<VecDataArray<double, 3>> normalsPtr    = std::make_shared<VecDataArray<double, 3>>(numVertices);
     VecDataArray<double, 3>&                 normals       = *normalsPtr;
-    std::shared_ptr<VecDataArray<double, 3>> tangentsPtr   = std::make_shared<VecDataArray<double, 3>>(numVertices);
-    VecDataArray<double, 3>&                 tangents      = *tangentsPtr;
+    std::shared_ptr<VecDataArray<float, 3>> tangentsPtr   = std::make_shared<VecDataArray<float, 3>>(numVertices);
+    VecDataArray<float, 3>&                 tangents      = *tangentsPtr;
     std::shared_ptr<VecDataArray<double, 3>> bitangentsPtr = std::make_shared<VecDataArray<double, 3>>(numVertices);
     VecDataArray<double, 3>&                 bitangents    = *bitangentsPtr;
 
@@ -135,7 +135,7 @@ AssimpMeshIO::convertAssimpMesh(aiMesh* importedMesh)
             auto tangentX = importedMesh->mTangents[i].x;
             auto tangentY = importedMesh->mTangents[i].y;
             auto tangentZ = importedMesh->mTangents[i].z;
-            tangents[i] = Vec3d(tangentX, tangentY, tangentZ);
+            tangents[i] = Vec3f(tangentX, tangentY, tangentZ);
 
             auto bitangentX = importedMesh->mBitangents[i].x;
             auto bitangentY = importedMesh->mBitangents[i].y;
